Except the not optimal merging of circles probably done
I cleaned up the code (or suppressed warnings :P) and fixed the bugs I found.
One could improve the way the circles are merged, to merge the closest circles together.
An option for that would be brute-forcing all solutions and then using  the one with the smallest cumulated distance.

I may do this in the future, or not. Not quite sure yet. Maybe an auto-updater would be nice too, which checks the last artifact jar submitted to github *hmm*